made changes to Actor and Movie servce and entity classes
- code for population movie, actor, and movieCast table

+++ moovi/src/main/java/com/dbms/project/moovi/data/entity/Screen.java
@@ -16,7 +16,7 @@ public class Screen {
 	
 	@Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int screenId;
+    private long screenId;
 	
 	@ManyToOne()
 	@JsonIgnore
@@ -30,11 +30,11 @@ public class Screen {
 		super();
 	}
 
-	public int getScreenId() {
+	public long getScreenId() {
 		return screenId;
 	}
 
-	public void setScreenId(int screenId) {
+	public void setScreenId(long screenId) {
 		this.screenId = screenId;
 	}
